Update language-multi-selector.component.ts
In this refactored code, the resetAndEmit method takes an array of languages as a parameter. This array is emitted through the save event emitter, and then the searchFilter and selectedLanguages properties are reset. The closeModal and saveLanguages methods now call resetAndEmit with the appropriate parameters. This eliminates the duplicated code and makes your component more maintainable.